refactoring back into standard layout
