+++ src/web/layouts/Standard.stories.tsx
@@ -12,6 +12,7 @@ import { Interview } from '@root/fixtures/articles/Interview';
 import { Quiz } from '@root/fixtures/articles/Quiz';
 import { Recipe } from '@root/fixtures/articles/Recipe';
 import { Comment } from '@root/fixtures/articles/Comment';
+import { MatchReport } from '@root/fixtures/articles/MatchReport';
 
 import { NAV } from '@root/fixtures/NAV';
 
@@ -146,3 +147,9 @@ RecipeStory.story = {
         chromatic: { viewports: [375] },
     },
 };
+
+export const MatchReportStory = () => {
+    const ServerCAPI = convertToStandard(MatchReport);
+    return <HydratedLayout ServerCAPI={ServerCAPI} />;
+};
+MatchReportStory.story = { name: 'MatchReport' };
